Introduce VrpFileDataProvider
